implemented oauth with ssr
