+++ clinician-dashboard/app/auth/logout/route.tsx
@@ -1,16 +1,15 @@
-import { createRouteHandlerClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
 import { type NextRequest, NextResponse } from "next/server";
+import { createClient } from "@/utils/supabase/server";
 
 export async function POST(req: NextRequest) {
-  const supabase = createRouteHandlerClient({ cookies });
+  const cookieStore = cookies();
+  const supabase = createClient(cookieStore);
 
   // Check if we have a session
   const {
     data: { session },
   } = await supabase.auth.getSession();
-
-  console.log(session);
 
   if (session) {
     await supabase.auth.signOut();
