--- clinician-dashboard/lib/supabase-server.tsx
@@ -1,5 +1,0 @@
-import { cookies } from "next/headers";
-import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
-
-export const createServerClient = () =>
-  createServerComponentClient({ cookies });
